# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OutreachSettingsBody": {
        "properties": {
          "allowed_categories": {
            "default": [
              "acknowledgement",
              "request_for_resume",
              "request_for_portfolio"
            ],
            "items": {
              "type": "string"
            },
            "title": "Allowed Categories",
            "type": "array"
          },
          "allowed_contact_types": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Allowed Contact Types",
            "type": "array"
          },
          "auto_reply_enabled": {
            "default": false,
            "title": "Auto Reply Enabled",
            "type": "boolean"
          },
          "auto_send_enabled": {
            "default": false,
            "title": "Auto Send Enabled",
            "type": "boolean"
          },
          "maximum_auto_replies_per_day": {
            "default": 5,
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Maximum Auto Replies Per Day",
            "type": "integer"
          },
          "maximum_daily_emails": {
            "default": 5,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Maximum Daily Emails",
            "type": "integer"
          },
          "minimum_classification_confidence": {
            "default": 0.92,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Minimum Classification Confidence",
            "type": "number"
          },
          "minimum_fit_score": {
            "default": 90,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Minimum Fit Score",
            "type": "integer"
          }
        },
        "title": "OutreachSettingsBody",
        "type": "object"
      },
...
      "SettingsBody": {
        "properties": {
          "career_preferences": {
            "additionalProperties": true,
            "title": "Career Preferences",
            "type": "object"
          },
          "outreach": {
            "$ref": "#/components/schemas/OutreachSettingsBody"
          }
        },
        "title": "SettingsBody",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Stable v1 backend for authenticated resume intelligence, career discovery, Gmail outreach, reply synchronization, and application tracking.",
    "title": "Chably AI - Resume Intelligence API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/account": {
      "delete": {
        "operationId": "delete_account_api_v1_account_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Account",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/account/export": {
      "get": {
        "operationId": "export_account_api_v1_account_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Account",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/applications": {
      "get": {
        "operationId": "applications_api_v1_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Applications",
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/v1/opportunities": {
      "get": {
        "operationId": "opportunity_list_api_v1_opportunities_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "minimum_fit",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100,
              "minimum": 0,
              "title": "Minimum Fit",
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Opportunity List",
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/v1/outreach": {
      "get": {
        "operationId": "outreach_list_api_v1_outreach_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Outreach List",
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/v1/saved-jobs": {
      "get": {
        "operationId": "saved_jobs_api_v1_saved_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "",
              "title": "Status",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Saved Jobs",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/search-history": {
      "get": {
        "operationId": "search_history_api_v1_search_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search History",
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/settings": {
      "get": {
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Settings",
        "tags": [
          "account"
        ]
      },
      "patch": {
        "operationId": "patch_settings_api_v1_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Settings",
        "tags": [
          "account"
        ]
      }
    },
...
  }
}